requirements.txt
fastapi
uvicorn[standard]
anthropic
PyGithub
python-dotenv
pydantic
httpx
cryptography
